Discord interactions refactor

+++ src/birthday/birthday.module.ts
@@ -3,25 +3,33 @@ import {
   getServerAnimeChannel,
   getServerBirthdayChannel,
   setServerBirthdayChannel,
-  updateServerLastWishes
-} from "../bot/bot.controller";
-import { checkAdmin, snowflakeToDate, stringReplacer } from "../helper/common";
-import Logger from "../helper/logger";
-import { getGuilds, setCommandExecutedCallback } from "../state/actions";
-import messageList from "../helper/messages";
-import { createInteractionResponse, sendMessage } from "../discord/rest";
+  updateServerLastWishes,
+} from '../bot/bot.controller';
+import { checkAdmin, snowflakeToDate, stringReplacer } from '../helper/common';
+import Logger from '../helper/logger';
+import {
+  getApplication,
+  getGuilds,
+  setCommandExecutedCallback,
+} from '../state/actions';
+import messageList from '../helper/messages';
+import {
+  createInteractionResponse,
+  editOriginalInteractionResponse,
+  sendMessage,
+} from '../discord/rest';
 import {
   addBirthday,
   getBirthdays,
   getUserBirthday,
   getBirthdaysByMonth,
-  updateLastWishes
-} from "./birthday.controller";
-import { IBirthday } from "./birthday.model";
-import { getOption, getOptionValue } from "../helper/modules.helper";
-import { interaction_response_type, no_mentions } from "../helper/constants";
+  updateLastWishes,
+} from './birthday.controller';
+import { IBirthday } from './birthday.model';
+import { getOption, getOptionValue } from '../helper/modules.helper';
+import { interaction_response_type, no_mentions } from '../helper/constants';
 
-const _logger = new Logger("birthday");
+const _logger = new Logger('birthday');
 let firstSetup: boolean = true;
 let checkTimeout: NodeJS.Timeout | null = null;
 
@@ -29,46 +37,40 @@ const handleChannelCommand = async (
   data: discord.interaction,
   option: discord.application_command_interaction_data_option
 ): Promise<void> => {
-  if (!checkAdmin(data.guild_id, data.member)) {
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.common.no_permission
-      }
-    });
-    return;
-  }
+  const app = getApplication();
+  if (app) {
+    if (!checkAdmin(data.guild_id, data.member)) {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.common.no_permission,
+      });
+      return;
+    }
 
-  const channel = getOptionValue<string>(option.options, "channel");
+    const channel = getOptionValue<string>(option.options, 'channel');
 
-  if (channel) {
-    await setServerBirthdayChannel(data.guild_id, channel);
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
+    if (channel) {
+      await setServerBirthdayChannel(data.guild_id, channel);
+      await editOriginalInteractionResponse(app.id, data.token, {
         content: stringReplacer(messageList.birthday.channel_set_success, {
-          channel: `<#${channel}>`
-        })
-      }
-    });
+          channel: `<#${channel}>`,
+        }),
+      });
 
-    _logger.log(
-      `Set anime channel to ${channel} in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
-    );
-  } else {
-    const ch = await getServerAnimeChannel(data.guild_id);
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
+      _logger.log(
+        `Set anime channel to ${channel} in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+      );
+    } else {
+      const ch = await getServerAnimeChannel(data.guild_id);
+      await editOriginalInteractionResponse(app.id, data.token, {
         content: stringReplacer(messageList.birthday.servers_channel, {
-          channel: `<#${ch}>`
-        })
-      }
-    });
+          channel: `<#${ch}>`,
+        }),
+      });
 
-    _logger.log(
-      `Get anime channel in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
-    );
+      _logger.log(
+        `Get anime channel in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+      );
+    }
   }
 };
 
@@ -76,47 +78,41 @@ const handleAddCommand = async (
   data: discord.interaction,
   option: discord.application_command_interaction_data_option
 ): Promise<void> => {
-  if (!checkAdmin(data.guild_id, data.member)) {
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.common.no_permission
-      }
-    });
-    return;
-  }
+  const app = getApplication();
+  if (app) {
+    if (!checkAdmin(data.guild_id, data.member)) {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.common.no_permission,
+      });
+      return;
+    }
 
-  const user = getOptionValue<string>(option.options, "user");
-  const day = getOptionValue<number>(option.options, "day");
-  const month = getOptionValue<number>(option.options, "month");
-  const year = getOptionValue<number>(option.options, "year");
+    const user = getOptionValue<string>(option.options, 'user');
+    const day = getOptionValue<number>(option.options, 'day');
+    const month = getOptionValue<number>(option.options, 'month');
+    const year = getOptionValue<number>(option.options, 'year');
 
-  const bd = await getUserBirthday(data.guild_id, user!);
+    const bd = await getUserBirthday(data.guild_id, user!);
 
-  if (bd) {
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.birthday.already_set
-      }
-    });
-  } else {
-    await addBirthday(data.guild_id, user!, day!, month!, year);
+    if (bd) {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.birthday.already_set,
+      });
+    } else {
+      await addBirthday(data.guild_id, user!, day!, month!, year);
 
-    const birthdayString = `${day}/${month}${year ? "/" + year : ""}`;
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
+      const birthdayString = `${day}/${month}${year ? '/' + year : ''}`;
+      await editOriginalInteractionResponse(app.id, data.token, {
         content: stringReplacer(messageList.birthday.set_success, {
           user: `<@${user}>`,
-          date: birthdayString
-        })
-      }
-    });
+          date: birthdayString,
+        }),
+      });
 
-    _logger.log(
-      `Add user ${user} birthday to ${birthdayString} in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
-    );
+      _logger.log(
+        `Add user ${user} birthday to ${birthdayString} in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+      );
+    }
   }
 };
 
@@ -124,39 +120,33 @@ const handleRemoveCommand = async (
   data: discord.interaction,
   option: discord.application_command_interaction_data_option
 ): Promise<void> => {
-  if (!checkAdmin(data.guild_id, data.member)) {
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.common.no_permission
-      }
-    });
-    return;
-  }
+  const app = getApplication();
+  if (app) {
+    if (!checkAdmin(data.guild_id, data.member)) {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.common.no_permission,
+      });
+      return;
+    }
 
-  const user = getOptionValue<string>(option.options, "user");
+    const user = getOptionValue<string>(option.options, 'user');
 
-  const bd = await getUserBirthday(data.guild_id, user!);
+    const bd = await getUserBirthday(data.guild_id, user!);
 
-  if (bd) {
-    await bd.delete();
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.birthday.remove_success
-      }
-    });
+    if (bd) {
+      await bd.delete();
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.birthday.remove_success,
+      });
 
-    _logger.log(
-      `Removed user ${user} birthday in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
-    );
-  } else {
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.birthday.not_found
-      }
-    });
+      _logger.log(
+        `Removed user ${user} birthday in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+      );
+    } else {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.birthday.not_found,
+      });
+    }
   }
 };
 
@@ -164,144 +154,136 @@ const handleGetCommand = async (
   data: discord.interaction,
   option: discord.application_command_interaction_data_option
 ): Promise<void> => {
-  const user = getOptionValue<string>(option.options, "user");
-  const month = getOptionValue<number>(option.options, "month");
+  const app = getApplication();
+  if (app) {
+    const user = getOptionValue<string>(option.options, 'user');
+    const month = getOptionValue<number>(option.options, 'month');
 
-  // no permission
-  if ((user || month) && !checkAdmin(data.guild_id, data.member)) {
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: messageList.common.no_permission
-      }
-    });
-    return;
-  }
+    // no permission
+    if ((user || month) && !checkAdmin(data.guild_id, data.member)) {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.common.no_permission,
+      });
+      return;
+    }
 
-  // for a user
-  if (user) {
-    const bd = await getUserBirthday(data.guild_id, user);
+    // for a user
+    if (user) {
+      const bd = await getUserBirthday(data.guild_id, user);
 
-    if (bd) {
-      const birthdayString = `${bd.day}/${bd.month}${
-        bd.year ? "/" + bd.year : ""
-      }`;
-      await createInteractionResponse(data.id, data.token, {
-        type: interaction_response_type.channel_message_with_source,
-        data: {
+      if (bd) {
+        const birthdayString = `${bd.day}/${bd.month}${
+          bd.year ? '/' + bd.year : ''
+        }`;
+        await editOriginalInteractionResponse(app.id, data.token, {
           content: stringReplacer(messageList.birthday.user, {
             user: `<@${user}>`,
-            date: birthdayString
+            date: birthdayString,
           }),
-          allowed_mentions: no_mentions
+          allowed_mentions: no_mentions,
+        });
+
+        _logger.log(
+          `Birthday for ${user} requested in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+        );
+      } else {
+        await editOriginalInteractionResponse(app.id, data.token, {
+          content: messageList.birthday.not_found,
+        });
+      }
+    }
+    // for a month
+    else if (month) {
+      const bd = await getBirthdaysByMonth(data.guild_id, month);
+
+      let message = '';
+
+      for (const b of bd) {
+        message += `<@${b.user}> - ${b.day}/${b.month}`;
+        if (b.year) {
+          message += `/${b.year}`;
         }
+
+        message += '\n';
+      }
+
+      if (message === '') {
+        message = stringReplacer(messageList.birthday.found_zero, {
+          month: month,
+        });
+      }
+
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: message,
+        allowed_mentions: no_mentions,
       });
 
       _logger.log(
-        `Birthday for ${user} requested in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+        `Birthday for month ${month} requested in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
       );
-    } else {
-      await createInteractionResponse(data.id, data.token, {
-        type: interaction_response_type.channel_message_with_source,
-        data: {
-          content: messageList.birthday.not_found
-        }
-      });
     }
-  }
-  // for a month
-  else if (month) {
-    const bd = await getBirthdaysByMonth(data.guild_id, month);
+    //for the user
+    else {
+      const bd = await getUserBirthday(data.guild_id, data.member.user!.id);
 
-    let message = "";
-
-    for (const b of bd) {
-      message += `<@${b.user}> - ${b.day}/${b.month}`;
-      if (b.year) {
-        message += `/${b.year}`;
-      }
-
-      message += "\n";
-    }
-
-    if (message === "") {
-      message = stringReplacer(messageList.birthday.found_zero, {
-        month: month
-      });
-    }
-
-    await createInteractionResponse(data.id, data.token, {
-      type: interaction_response_type.channel_message_with_source,
-      data: {
-        content: message,
-        allowed_mentions: no_mentions
-      }
-    });
-
-    _logger.log(
-      `Birthday for month ${month} requested in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
-    );
-  }
-  //for the user
-  else {
-    const bd = await getUserBirthday(data.guild_id, data.member.user!.id);
-
-    if (bd) {
-      const birthdayString = `${bd.day}/${bd.month}${
-        bd.year ? "/" + bd.year : ""
-      }`;
-      await createInteractionResponse(data.id, data.token, {
-        type: interaction_response_type.channel_message_with_source,
-        data: {
+      if (bd) {
+        const birthdayString = `${bd.day}/${bd.month}${
+          bd.year ? '/' + bd.year : ''
+        }`;
+        await editOriginalInteractionResponse(app.id, data.token, {
           content: stringReplacer(messageList.birthday.user, {
             user: `<@${data.member.user!.id}>`,
-            date: birthdayString
+            date: birthdayString,
           }),
-          allowed_mentions: no_mentions
-        }
-      });
+          allowed_mentions: no_mentions,
+        });
 
-      _logger.log(
-        `Birthday requested in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
-      );
-    } else {
-      await createInteractionResponse(data.id, data.token, {
-        type: interaction_response_type.channel_message_with_source,
-        data: {
-          content: messageList.birthday.not_found
-        }
-      });
+        _logger.log(
+          `Birthday requested in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+        );
+      } else {
+        await editOriginalInteractionResponse(app.id, data.token, {
+          content: messageList.birthday.not_found,
+        });
+      }
     }
   }
 };
 
 const handleServerCommand = async (
-  data: discord.interaction,
-  option: discord.application_command_interaction_data_option
+  data: discord.interaction
 ): Promise<void> => {
-  const serverDate = snowflakeToDate(data.guild_id);
+  const app = getApplication();
+  if (app) {
+    const serverDate = snowflakeToDate(data.guild_id);
 
-  const birthdayString = `${serverDate.getDate()}/${
-    serverDate.getMonth() + 1
-  }/${serverDate.getFullYear()}`;
+    const birthdayString = `${serverDate.getDate()}/${
+      serverDate.getMonth() + 1
+    }/${serverDate.getFullYear()}`;
 
-  await createInteractionResponse(data.id, data.token, {
-    type: interaction_response_type.channel_message_with_source,
-    data: {
+    await editOriginalInteractionResponse(app.id, data.token, {
       content: stringReplacer(messageList.birthday.server, {
-        date: birthdayString
-      })
-    }
-  });
+        date: birthdayString,
+      }),
+    });
+
+    _logger.log(
+      `Get server birthday date in ${data.guild_id} by ${data.member.user?.username}#${data.member.user?.discriminator}`
+    );
+  }
 };
 
 const commandExecuted = async (data: discord.interaction): Promise<void> => {
-  if (data.data && data.data.name === "birthday" && data.data.options) {
-    const channel = getOption(data.data.options, "channel");
-    const add = getOption(data.data.options, "add");
-    const remove = getOption(data.data.options, "remove");
-    const get = getOption(data.data.options, "get");
-    const server = getOption(data.data.options, "server");
+  if (data.data && data.data.name === 'birthday' && data.data.options) {
+    await createInteractionResponse(data.id, data.token, {
+      type: interaction_response_type.acknowledge_with_source,
+    });
+
+    const channel = getOption(data.data.options, 'channel');
+    const add = getOption(data.data.options, 'add');
+    const remove = getOption(data.data.options, 'remove');
+    const get = getOption(data.data.options, 'get');
+    const server = getOption(data.data.options, 'server');
 
     if (channel) {
       return await handleChannelCommand(data, channel);
@@ -320,20 +302,27 @@ const commandExecuted = async (data: discord.interaction): Promise<void> => {
     }
 
     if (server) {
-      return handleServerCommand(data, server);
+      return handleServerCommand(data);
     }
 
     _logger.error(
-      "UNKNOWN COMMAND",
+      'UNKNOWN COMMAND',
       data.data.options[0].name,
       data.data.options[0].options,
       data.data.options[0].value
     );
+
+    const app = getApplication();
+    if (app) {
+      await editOriginalInteractionResponse(app.id, data.token, {
+        content: messageList.common.internal_error,
+      });
+    }
   }
 };
 
 const checkBirthdays = async () => {
-  _logger.log("Checking birthdays");
+  _logger.log('Checking birthdays');
 
   if (checkTimeout) {
     clearTimeout(checkTimeout);
@@ -366,7 +355,7 @@ const checkBirthdays = async () => {
       if (!lastWishes || lastWishes < year) {
         const message = stringReplacer(messageList.birthday.server_bday, {
           age: (year - serverDate.getFullYear()).toString(),
-          name: s.name
+          name: s.name,
         });
 
         await sendMessage(serverChannels[s.id]!, message);
@@ -416,7 +405,7 @@ const checkBirthdays = async () => {
   const time =
     1000 -
     milliseconds + //ms to next s
-    (60 - seconds - 1) * 1000 + //s to next m
+    (60 - seconds) * 1000 + //s to next m
     (60 - minutes - 1) * 60 * 1000 + //m to next h
     ((hours > 12 ? 24 : 12) - hours - 1) * 60 * 60 * 1000; //h to next 12/24
